+++ src/objects/Audio.js
@@ -20,6 +20,7 @@ class Audio extends GameObject {
 	register(scene) {
 		const { name } = this.props;
 		this.scene = scene;
+		this.preRegister();
 		this.instance = scene.sound.add(name);
 		this.registered = true;
 		this.update(this.props);