+++ src/objects/Container.js
@@ -10,6 +10,7 @@ class Container extends GameObject {
 	register(scene) {
 		const { x, y } = this.props;
 		this.scene = scene;
+		this.preRegister();
 		this.instance = new Phaser.GameObjects.Container(scene, x, y);
 		scene.add.displayList.add(this.instance);
 		this.registered = true;