+++ src/objects/GameObject/index.js
@@ -52,6 +52,11 @@ class GameObject {
 			...this.eventMap
 		};
 
+		this.allowedDefaultProps = {
+			...pick(defaultProps, this.allowedProps),
+			...this.defaultProps
+		};
+
 		this.eventNames = Object.keys(this.fullEventMap);
 	}
 
@@ -118,7 +123,10 @@ class GameObject {
 
 	/* eslint-disable complexity */
 	update(newProps, oldProps) {
-		this.props = newProps;
+		this.props = {
+			...this.allowedDefaultProps,
+			...newProps
+		};
 
 		if (!this.registered) return;
 
@@ -133,8 +141,7 @@ class GameObject {
 		const { instance, fullEventMap, interactive } = this;
 
 		const props = {
-			...pick(defaultProps, this.allowedProps),
-			...this.defaultProps,
+			...this.allowedDefaultProps,
 			...pick(newProps, this.allowedProps),
 			...pick(newProps, this.eventNames)
 		};
@@ -270,6 +277,7 @@ Object.assign(GameObject.prototype, {
 	performedProps,
 	allowedProps,
 	transitionProps,
+	preRegister: noop,
 	postRegister: noop,
 	defaultProps: emptyObject,
 	eventMap: emptyObject