+++ src/objects/Polygon.js
@@ -40,6 +40,7 @@ class Polygon extends GameObject {
 	register(scene) {
 		const { x, y, points, fillColor } = this.props;
 		this.scene = scene;
+		this.preRegister();
 		this.instance = scene.add.polygon(x, y, points, fillColor);
 		this.registered = true;
 		this.update(this.props);