+++ src/objects/Rectangle.js
@@ -35,6 +35,7 @@ class Rectangle extends GameObject {
 	register(scene) {
 		const { x, y, width, height, fillColor } = this.props;
 		this.scene = scene;
+		this.preRegister();
 		this.instance = scene.add.rectangle(x, y, width, height, fillColor);
 		this.registered = true;
 		this.update(this.props);