+++ src/objects/Sprite.js
@@ -54,6 +54,7 @@ class Sprite extends GameObject {
 	register(scene) {
 		const { x, y, texture, frame } = this.props;
 		this.scene = scene;
+		this.preRegister();
 		this.instance = new Phaser.GameObjects.Sprite(scene, x, y, texture, frame);
 		this.registered = true;
 		scene.add.displayList.add(this.instance);