+++ src/objects/Star.js
@@ -36,6 +36,7 @@ class Star extends GameObject {
 	register(scene) {
 		const { x, y, points, innerRadius, outerRadius, fillColor } = this.props;
 		this.scene = scene;
+		this.preRegister();
 		this.instance = scene.add.star(x, y, points, innerRadius, outerRadius, fillColor);
 		this.registered = true;
 		this.update(this.props);